Add check for .dockerignore

+++ entrypoint.sh
@@ -2,7 +2,7 @@
 
 image=${image:-alpine}
 
-if [ "$1" = "-f" ] || ([ ! -f '/skel/Dockerfile' ] && [ ! -f '/skel/entrypoint.sh' ]);
+if [ "$1" = "-f" ] || ([ ! -f '/skel/Dockerfile' ] && [ ! -f '/skel/entrypoint.sh' ] && [ ! -f '/skel/.dockerignore' ]);
 then
     sed "s#__IMAGE__#${image}#" /skeleton/Dockerfile > /skel/Dockerfile
     cp /skeleton/entrypoint.sh /skel
